Add java layer to webtesting container. (#52)
In any Bazel project that has any test target, the test executor java
tool will need to run which needs a Java toolchain. If we do not install
a JVM in the container, and do not specify the --host_javabase and
--javabase flags, Bazel will download download remote JDK according to
the host machine and os type, this is not guaranteed to work in
cross-platform build/test cases (mac host --> linux exec). So we still
need to install our controlled Java in the container and specify java
flags to point to it.