Auto-tag bazel container and bazel docker gcloud container as latest. (#496)
For bazel container, it should be ok to just tag the new one as latest
directly. Although the marketplace syncing bot might sync the container
while some other container generation jobs that are using
l.gcr.io/google/bazel container are still running, but we can still tell
exactly which Bazel version the jobs are using from the logs.

For the bazel docker gcloud container, since we no longer use this one
as the base container to generate other containers, we can safely tag it
without affecting the generation of other containers.